backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.119.1",
    "langchain>=1.0.1",
    "langchain-openai>=1.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]
